// src/api/generated/schemas/userDto.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * TripBite API
 * OpenAPI spec version: 1.0
 */
import type { RegionCode } from './regionCode';
import type { TravelTypeBriefDto } from './travelTypeBriefDto';

export interface UserDto {
  homeRegion: RegionCode;
  id: string;
  username: string;
  nickname: string;
  email: string;
  isOnboarded: boolean;
  /** @nullable */
  avatarUrl: string | null;
  /** @nullable */
  travelType: TravelTypeBriefDto | null;
}